Removed queue from ButtonVariable. Added protection for HOVERED state in changeState.
A queue in ButtonVariable would essentially require the event system to be rebuilt as a queue as well, and tests on the original engine show that events are emitted immediately rather than at the end of the renderer loop. At this moment, button states are changed immediately, and protection have been added for HOVERED state so that it does not emit ONFOCUSON event if, for example, key release events would disable the button in the meantime. Fixes PALISADA13 in Reksio i Skarb Piratów.