Migration from Music to Sound, ONFINISHED and SETFREQ were implemented.
The migration was necessary because Music does not support pitch shifting as the data is streamed from disk. On the other hand, in Sound we do not have any indicator of the audio position or the listener when the audio stops playing. Therefore, it was necessary to read the duration from the header of the audio file and in each frame it checks if the audio has finished playing.