fix: Fixed infinite loop when trying to get variable from class instance.
For now, I'm fixing Reksio i Czarodzieje. README needs to be updated. Added protection for arguments as it must be primitive value, any other types are illegal. Constructor as first parameter takes name of class instance. But getVariable resolves it as InstanceVariable, so it caused adding context as classContext and infinite loop.