feat: Working on FOLLOWPATH
I have problem with this. In assembly, there's an infinite loop interrupted by an empty queue, but in ARCADE, in Reksio i Czarodziej, the FOLLOWPATH method is called from STL0:ONFRAMECHANGED^ELAPSE and also returns the distance to the next waypoint. In this form, it would be code that blocks the main thread. Analysis is ongoing.